% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLookUp.R
\name{batchLookUp}
\alias{batchLookUp}
\title{Combines results of multiple screens in single dataframe}
\usage{
batchLookUp(x, nameOutput, interest, genelist, threshold = 100)
}
\arguments{
\item{x}{A dataframe containing info on each screen that needs to be added
to the final dataframe.}

\item{nameOutput}{Name of the output dataframe that will added to your global
environment.}

\item{interest}{Indicates in which batchScreenAnalysis column you are
interested in. Chose between "ratio" and "MEDnorm".}

\item{genelist}{List of all genes you want to have in you final data frame.}

\item{threshold}{Selects minimal pixels size for first media. Default is 100}
}
\description{
Based on a sample table, extracts results of screens out of the
    .txt file generated by batchScreenAnalysis and combines them into a
    single datafame.
}
\examples{
    # Define sample table
    location_results <- c("/Desktop/screen1","/Desktop/screen2","/Desktop/screen3")
    screen_name <- c("screen1","screen2","screen3")
    type <- c("mat","mat","tel")
    mutant <- c("wt","abc","wt")
    assay_stress <- c("std","std","temp")
    media1 <- c("NS","NS","NS")
    media2 <- c("URA","URA","URA")
    sampleTable <- data.frame(location_results, screen_name, type, mutant, assay_stress, media1, media2)
    # Genes you are interested in
    genelist <- c("abc","def","ghi", "jkl","mno","oqr")
    # Call the function
    batchLookUp(sampleTable, "screen.df", "ratio", genelist, threshold = 150)

}
