% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNamer.R
\name{batchNamer}
\alias{batchNamer}
\title{Renames files as they come from the colonie size analyzer software}
\usage{
batchNamer(x, y)
}
\arguments{
\item{x}{}
}
\description{
Renames file that are produced by colonie size analyzer software so the names are compatible with sgaMethods
}
\examples{
batchNamer(screenTable, y)

}
