% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScreenAnalysis.R
\name{batchScreenAnalysis}
\alias{batchScreenAnalysis}
\title{Analyzes a single screen}
\usage{
batchScreenAnalysis(x, borderEffectCorrection, threshold = 100,
  plateMED_med2_limit = 2)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
batch.screen.analysis(screenTable)

}
