% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{Makes a correlation matrix of a data frame with screen results}
\usage{
corMatrix(x)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
corMatrix(screenTable)

}
