% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.n.dirtyPCA.R
\name{quick.n.dirtyPCA}
\alias{quick.n.dirtyPCA}
\title{Makes a simple PCA plot of dataframe containing screen results}
\usage{
quick.n.dirtyPCA(x)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
quick.n.dirty(screenTable)

}
