% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgaMethodsCorMatrix.R
\name{sgaMethodsCorMatrix}
\alias{sgaMethodsCorMatrix}
\title{Makes a correlation matrix of a data frame with screen results}
\usage{
sgaMethodsCorMatrix(x)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
sgaMethodsCorMatrix(screenTable)

}
