% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgaMethodsPCA.R
\name{sgaMethodsPCA}
\alias{sgaMethodsPCA}
\title{Makes a simple PCA plot of dataframe containing screen results}
\usage{
sgaMethodsPCA(x)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
sgaMethodsPCA(screenTable)

}
