% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Reorders the dataframe containing your screen results}
\usage{
shuffle(x, order.df, to.order, nameOutput)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
shuffle(screen.df, order.df, to.order, nameOutput)

}
