% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNamer.R
\name{batchNamer}
\alias{batchNamer}
\title{Renames files as they come from the colonie size analyzer}
\usage{
batchNamer(x, y)
}
\arguments{
\item{x}{}
}
\description{
Analyzes SGAs
}
\examples{
batchNamer(screenTable, y)

}
